修改生成信息

+++ src/main/java/com/oner365/sys/controller/auth/AuthController.java
@@ -184,8 +184,10 @@ public class AuthController extends BaseController {
   @ApiOperationSupport(order = 5)
   @PostMapping("/logout")
   public String logout(@ApiIgnore @CurrentUser AuthUser authUser) {
-    String key = CacheConstants.CACHE_LOGIN_NAME + authUser.getUserName();
-    redisCache.deleteObject(key);
+    if (authUser != null) {
+      String key = CacheConstants.CACHE_LOGIN_NAME + authUser.getUserName();
+      redisCache.deleteObject(key);
+    }
     return ResultEnum.SUCCESS.getName();
   }
